Fixed a number of issues in Grip
The tool was obviously built to be a Web server first and command line tool second (the export feature was more as an afterthought).  The command line behavior on errors is bad - it just produces a web server error page and doesn't exit with a failure code.  It exits with a success code!  I have patched the tool to now exit with error code 1 when file is missing and when decode errors happen.  I looked at ways to make those changes only when operating in export mode but the architecture wasn't designed for this so that would be major surgery.  Since we will only be using the export feature I made the changes with that in mind.